requirements.txt
amqp==5.3.1
attrs==25.1.0
BareNecessities==0.2.8
bcrypt==4.0.1
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
cachelib==0.9.0
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cors==1.0.1
coverage==7.7.0
cryptography==44.0.2
Deprecated==1.2.18
dnspython==2.7.0
dotenv==0.9.9
-e git+https://github.com/Onnboy/Disparo-de-Email-Marketing.git@44508059a5d676847ed96f38eb1cb3c3832d181b#egg=email_sender
filelock==3.17.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-Caching==2.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.5.0
Flask-Mail==0.10.0
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
kombu==5.5.2
limiter==0.5.0
limits==3.9.0
Mail==2.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.2
mysql-connector==2.2.9
mysql-connector-python==9.2.0
numpy==2.2.4
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
pycparser==2.22
Pygments==2.19.1
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
requests-file==2.1.0
resolv==1.3.0
resolver==0.2.1
rich==13.9.4
rpds-py==0.23.1
setuptools==75.8.0
sib-api-v3-sdk==7.6.0
six==1.17.0
soupsieve==2.6
starkbank-ecdsa==2.2.0
StrEnum==0.4.15
tldextract==5.1.3
token-bucket==0.3.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
